put back notebook, remove processing examples
